Review comment fixes (code restructuring)
