+++ presto-native-execution/presto_cpp/main/SystemConnector.h
@@ -184,7 +184,9 @@ class SystemPrestoToVeloxConnector final : public PrestoToVeloxConnector {
 
   std::unique_ptr<velox::connector::ConnectorSplit> toVeloxSplit(
       const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit) const final;
+      const protocol::ConnectorSplit* connectorSplit,
+      const std::map<std::string, std::string>& extraCredentials = {})
+      const final;
 
   std::unique_ptr<velox::connector::ColumnHandle> toVeloxColumnHandle(
       const protocol::ColumnHandle* column,
