+++ presto-native-execution/presto_cpp/main/types/PrestoToVeloxConnector.h
@@ -45,7 +45,9 @@ class PrestoToVeloxConnector {
   [[nodiscard]] virtual std::unique_ptr<velox::connector::ConnectorSplit>
   toVeloxSplit(
       const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit) const = 0;
+      const protocol::ConnectorSplit* connectorSplit,
+      const std::map<std::string, std::string>& extraCredentials = {})
+      const = 0;
 
   [[nodiscard]] virtual std::unique_ptr<velox::connector::ColumnHandle>
   toVeloxColumnHandle(
@@ -115,7 +117,9 @@ class HivePrestoToVeloxConnector final : public PrestoToVeloxConnector {
 
   std::unique_ptr<velox::connector::ConnectorSplit> toVeloxSplit(
       const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit) const final;
+      const protocol::ConnectorSplit* connectorSplit,
+      const std::map<std::string, std::string>& extraCredentials = {})
+      const final;
 
   std::unique_ptr<velox::connector::ColumnHandle> toVeloxColumnHandle(
       const protocol::ColumnHandle* column,
@@ -166,7 +170,9 @@ class IcebergPrestoToVeloxConnector final : public PrestoToVeloxConnector {
 
   std::unique_ptr<velox::connector::ConnectorSplit> toVeloxSplit(
       const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit) const final;
+      const protocol::ConnectorSplit* connectorSplit,
+      const std::map<std::string, std::string>& extraCredentials = {})
+      const final;
 
   std::unique_ptr<velox::connector::ColumnHandle> toVeloxColumnHandle(
       const protocol::ColumnHandle* column,
@@ -192,7 +198,9 @@ class TpchPrestoToVeloxConnector final : public PrestoToVeloxConnector {
 
   std::unique_ptr<velox::connector::ConnectorSplit> toVeloxSplit(
       const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit) const final;
+      const protocol::ConnectorSplit* connectorSplit,
+      const std::map<std::string, std::string>& extraCredentials = {})
+      const final;
 
   std::unique_ptr<velox::connector::ColumnHandle> toVeloxColumnHandle(
       const protocol::ColumnHandle* column,
@@ -210,32 +218,4 @@ class TpchPrestoToVeloxConnector final : public PrestoToVeloxConnector {
   std::unique_ptr<protocol::ConnectorProtocol> createConnectorProtocol()
       const final;
 };
-
-#ifdef PRESTO_ENABLE_ARROW_FLIGHT_CONNECTOR
-class ArrowPrestoToVeloxConnector final : public PrestoToVeloxConnector {
- public:
-  explicit ArrowPrestoToVeloxConnector(std::string connectorName)
-      : PrestoToVeloxConnector(std::move(connectorName)) {}
-
-  std::unique_ptr<velox::connector::ConnectorSplit> toVeloxSplit(
-      const protocol::ConnectorId& catalogId,
-      const protocol::ConnectorSplit* connectorSplit) const final;
-
-  std::unique_ptr<velox::connector::ColumnHandle> toVeloxColumnHandle(
-      const protocol::ColumnHandle* column,
-      const TypeParser& typeParser) const final;
-
-  std::unique_ptr<velox::connector::ConnectorTableHandle> toVeloxTableHandle(
-      const protocol::TableHandle& tableHandle,
-      const VeloxExprConverter& exprConverter,
-      const TypeParser& typeParser,
-      std::unordered_map<
-          std::string,
-          std::shared_ptr<velox::connector::ColumnHandle>>& assignments)
-      const final;
-
-  std::unique_ptr<protocol::ConnectorProtocol> createConnectorProtocol()
-      const final;
-};
-#endif
 } // namespace facebook::presto
